refactor: move prependPendingPeerMessage to its own file
