+++ src/game-logic/reducers/prependPendingPeerMessage.js
@@ -1,0 +1,21 @@
+import { MAX_PENDING_PEER_MESSAGES } from '../../constants'
+
+/**
+ * @param {farmhand.state} state
+ * @param {string} peerMessage
+ * @param {string?} [severity='info']
+ * @returns {farmhand.state}
+ */
+export const prependPendingPeerMessage = (
+  state,
+  message,
+  severity = 'info'
+) => {
+  return {
+    ...state,
+    pendingPeerMessages: [
+      { id: state.id, message, severity },
+      ...state.pendingPeerMessages,
+    ].slice(0, MAX_PENDING_PEER_MESSAGES),
+  }
+}
