+++ src/game-logic/reducers/prependPendingPeerMessage.test.js
@@ -1,26 +1,10 @@
 import { MAX_PENDING_PEER_MESSAGES } from '../../constants'
 
-import * as fn from './'
-
-jest.mock('../../data/achievements')
-jest.mock('../../data/maps')
-jest.mock('../../data/items')
-jest.mock('../../data/levels', () => ({ levels: [], itemUnlockLevels: {} }))
-jest.mock('../../data/recipes')
-jest.mock('../../data/shop-inventory')
-jest.mock('../../utils/isRandomNumberLessThan')
-
-jest.mock('../../constants', () => ({
-  __esModule: true,
-  ...jest.requireActual('../../constants'),
-  COW_HUG_BENEFIT: 0.5,
-  CROW_CHANCE: 0,
-  PRECIPITATION_CHANCE: 0,
-}))
+import { prependPendingPeerMessage } from './prependPendingPeerMessage'
 
 describe('prependPendingPeerMessage', () => {
   test('prepends a message', () => {
-    const { pendingPeerMessages } = fn.prependPendingPeerMessage(
+    const { pendingPeerMessages } = prependPendingPeerMessage(
       { id: 'abc123', pendingPeerMessages: [] },
       'hello world'
     )
@@ -31,7 +15,7 @@ describe('prependPendingPeerMessage', () => {
   })
 
   test('limits the amount of stored messages', () => {
-    const { pendingPeerMessages } = fn.prependPendingPeerMessage(
+    const { pendingPeerMessages } = prependPendingPeerMessage(
       {
         id: 'abc123',
         pendingPeerMessages: new Array(50).fill({
